비즈니스 로직: 투표함 목록 및 상세 조회 API 구현 (NPE 방어 로직 포함)
안정성 리팩토링: GroupMember 연관관계의 부적절한 CascadeType.ALL을 제거하여 의도치 않은 데이터 삭제 방지
품질 보증: JUnit 5/Mockito를 활용하여 주요 서비스 로직에 대한 단위 테스트 100% 통과
협업 최적화: Swagger @Schema 적용 및 DTO 정리를 통한 API 명세 구체화

+++ src/main/java/com/example/anything/vote/internal/domain/VoteOption.java
@@ -26,7 +26,7 @@ public class VoteOption {
 
     private String menuName;
 
-    private Long count;
+    private Integer count;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "ballot_box_id")
@@ -41,12 +41,12 @@ public class VoteOption {
         voteOption.menuId = id;
         voteOption.menuName = menuName;
         voteOption.ballotBox = ballotBox;
-        voteOption.count = 0L;
+        voteOption.count = 0;
 
         return voteOption;
     }
 
     public void addCount() {
-        this.count = (this.count == null) ? 1L : this.count + 1;
+        this.count = (this.count == null) ? 1 : this.count + 1;
     }
 }
